Reads PS/2 interface correctly, with the occasional error.
Haven't tested writing yet. The errors seem to be on the stop bit?
The driver is currently configured so it won't request a resend, but
will instead passively continue recieving the packet, while reporting
the bit the most recent error occured on. Perhaps a more sophisticated
error logging system is needed to debug this, but since it works most
of the time, I'm satisfied for now.

Important changes in this commit are things like switching the data pins,
and switching from even parity to odd parity. Most of the changes in main.c
are superficial to the function of the driver.